Support for Additional File Extensions
This patch addresses the issue of saving files without an extension. Previously TEdit would append .txt to the end of filenames when saved without an extension, this is no longer the case allowing files to be saved sans extension. Additionally for Android API level 26, Oreo or 8.0, and higher TEdit will now use Android's built-in file type detector to filter supported files by their MIME type. Unfortunately, at least on my device, this still does not detect text files with no extension. In order to accomplish this TEdit would need to read a number of bytes from the file use that information to guess its MIME type which seems a bit like overkill right now. One could still select the un-extensioned file with a file manager and select TEdit as the application to open the file with.