Handling Secure Directories and Files
Some devices will return an empty file array when querying the contents of secure directories outside of public storage. This presents a problem when attempting to access the parent directory of the root public storage directory. You can navigate up to the parent directory, but no contents will be displayed so you cannot navigate back into public storage. When navigating to a parent directory TEdit now checks to see if the content array of the parent directory is empty or null, if so the browser will remain in the current directory and display a permission denied message.